+++ examples/basic.rs
@@ -29,6 +29,9 @@ fn main() {
     context.add("numbers", &vec![1,2,3]);
     context.add("bio", &"<script>alert('pwnd');</script>");
 
+    // A one off template
+    Tera::one_off("hello", Context::new(), true).unwrap();
+
     match TEMPLATES.render("users/profile.html", context) {
         Ok(s) => println!("{:?}", s),
         Err(e) => {