+++ src/context.rs
@@ -3,7 +3,6 @@ use std::collections::BTreeMap;
 use serde::ser::Serialize;
 use serde_json::value::{Value, to_value};
 
-pub type TemplateContext = BTreeMap<String, Value>;
 
 /// The struct that holds the context of a template rendering.
 ///